Add nlohmann/json library and a simple use example
The json library created by nlohmann aims to make JSON a "first-class
data type" in C++. Its intuitive interface and single-header
distribution make it a suitable choice for the project.

More info: https://github.com/nlohmann/json

This commit adds the library as a static dependency and uses it in a
simple test example.