Implement function for finding posTags
PosTag values are needed as arguments for the Polem lemmatization. The
function added in this commit finds the posTag labels in a provided
label array and returns a list of them that is sorted by the
"startToken" key. NER labels will be matched to posTags by the token
position. It is assumed that there is exactly one posTag for each token
position and that for each of them endToken - startToken == 1.