Implement lemmatization algorithm
The "lemmatizeNerLabels" function can now perform lemmatization of
provided NER labels if given corrent posTag and lemma label values. To
work correctly, the mentioned label values need to be transformed to
lower case strings, before passing to Polem. At this point the
transformation is performed ignoring any localization concerns. This
obviously would need to be addressed in a production-ready solution.