Integrate Pistache library
Pistache is a REST framework for modern C++. It offers a clean API and
has decent documentation. Should be a good pick for this task.

This commit includes Pistache in the project and implements a bare-bones
server and request handler. A Bash script is provided to test the sever
using curl tool.