Complete initial implementation of Polem lemmatization REST service
This is the first working version of the microservice.

The server accepts POST requests with "application/json" content type,
sent to local port 5000. The request body is expected to contain the
input JSON. The test script loads the JSON from the provided path.
Currently the maxium allowed size of the request is 1024*1024 bytes (1
MB).

The received JSON is processed using functions implemented in the
previous commits. One additional Polem label is added for each NER
label. The result JSON is placed in the HTML response body and sent with
"application/json" content type.